server/mqtt-client/requirements.txt
paho-mqtt
elasticsearch